feat: イベント処理ロジックを`events.rs`に集約し、各モードのキーイベントハンドリングを実装
