+++ src/main.rs
@@ -2,6 +2,7 @@ mod api_client;
 mod app;
 mod config;
 mod error;
+mod events;
 mod help;
 mod reports;
 mod stats;
@@ -10,24 +11,14 @@ mod ui;
 
 use crate::{
     api_client::{
-        format_evaluation_display, parse_evaluation, ApiClient, ApiClientLike, OverallEvaluation,
+        ApiClient, ApiClientLike, OverallEvaluation, format_evaluation_display, parse_evaluation,
     },
-    app::{App, MENU_OPTIONS, ViewMode},
+    app::{App, ViewMode},
     error::AppError,
+    events::AppAction,
     stats::EvaluationScores,
 };
-use rat_text::event::HandleEvent;
-use ratatui::{
-    crossterm::event::{self, Event, KeyCode, KeyEventKind, KeyModifiers},
-    widgets::{Paragraph, Wrap},
-};
-use std::{env, time::Duration};
-
-/// Event polling interval in milliseconds
-const EVENT_POLL_INTERVAL_MS: u64 = 100;
-
-/// Overlay size as percentage of screen
-const OVERLAY_SIZE_PERCENT: u16 = 75;
+use std::env;
 
 #[tokio::main]
 async fn main() -> Result<(), AppError> {
@@ -42,7 +33,7 @@ async fn main() -> Result<(), AppError> {
     while !app.should_quit {
         tui.draw(|frame| ui::render(&mut app, frame))?;
 
-        if let Some(result) = handle_events(&mut app).await? {
+        if let Some(result) = events::handle_events(&mut app).await? {
             match result {
                 AppAction::StartTraining => {
                     app.view_mode = ViewMode::Normal;
@@ -64,51 +55,47 @@ async fn main() -> Result<(), AppError> {
                             .evaluate_summary(app.original_text.clone(), summary)
                             .await
                         {
-                            Ok(evaluation) => {
-                                match parse_evaluation(&evaluation) {
-                                    Ok(parsed) => {
-                                        app.evaluation_passed =
-                                            matches!(parsed.overall, OverallEvaluation::Pass);
-                                        app.evaluation_text = format_evaluation_display(&parsed);
-                                        app.show_evaluation_overlay = true;
-                                        app.evaluation_overlay_scroll = 0;
-                                        app.is_evaluating = false;
-                                        app.status_message = "評価が完了しました。'e'で切替、'n'で次へ。"
-                                            .to_string();
+                            Ok(evaluation) => match parse_evaluation(&evaluation) {
+                                Ok(parsed) => {
+                                    app.evaluation_passed =
+                                        matches!(parsed.overall, OverallEvaluation::Pass);
+                                    app.evaluation_text = format_evaluation_display(&parsed);
+                                    app.show_evaluation_overlay = true;
+                                    app.evaluation_overlay_scroll = 0;
+                                    app.is_evaluating = false;
+                                    app.status_message =
+                                        "評価が完了しました。'e'で切替、'n'で次へ。".to_string();
 
-                                        let scores = EvaluationScores {
-                                            appropriate: parsed.appropriate,
-                                            importance: parsed.importance,
-                                            conciseness: parsed.conciseness,
-                                            accuracy: parsed.accuracy,
-                                            improvement1: parsed.improvement1,
-                                            improvement2: parsed.improvement2,
-                                            improvement3: parsed.improvement3,
-                                            overall_passed: app.evaluation_passed,
-                                        };
+                                    let scores = EvaluationScores {
+                                        appropriate: parsed.appropriate,
+                                        importance: parsed.importance,
+                                        conciseness: parsed.conciseness,
+                                        accuracy: parsed.accuracy,
+                                        improvement1: parsed.improvement1,
+                                        improvement2: parsed.improvement2,
+                                        improvement3: parsed.improvement3,
+                                        overall_passed: app.evaluation_passed,
+                                    };
 
-                                        app.stats.add_result_with_evaluation(
-                                            app.evaluation_passed,
-                                            Some(scores),
-                                        );
-                                        if let Err(e) = app.stats.save() {
-                                            app.status_message =
-                                                format!("警告: 統計の保存に失敗しました: {}", e);
-                                            eprintln!("Failed to save stats: {}", e);
-                                        }
-                                    }
-                                    Err(_) => {
-                                        app.evaluation_text =
-                                            "評価結果の形式が不正です".to_string();
-                                        app.evaluation_passed = false;
-                                        app.show_evaluation_overlay = true;
-                                        app.evaluation_overlay_scroll = 0;
-                                        app.is_evaluating = false;
+                                    app.stats.add_result_with_evaluation(
+                                        app.evaluation_passed,
+                                        Some(scores),
+                                    );
+                                    if let Err(e) = app.stats.save() {
                                         app.status_message =
-                                            "評価結果の形式が不正です".to_string();
+                                            format!("警告: 統計の保存に失敗しました: {}", e);
+                                        eprintln!("Failed to save stats: {}", e);
                                     }
                                 }
-                            }
+                                Err(_) => {
+                                    app.evaluation_text = "評価結果の形式が不正です".to_string();
+                                    app.evaluation_passed = false;
+                                    app.show_evaluation_overlay = true;
+                                    app.evaluation_overlay_scroll = 0;
+                                    app.is_evaluating = false;
+                                    app.status_message = "評価結果の形式が不正です".to_string();
+                                }
+                            },
                             Err(e) => {
                                 app.evaluation_text = format!("Error: {}", e);
                                 app.evaluation_passed = false;
@@ -139,226 +126,6 @@ async fn main() -> Result<(), AppError> {
 
     tui::restore()?;
     Ok(())
-}
-
-enum AppAction {
-    Evaluate,
-    NextTraining,
-    StartTraining,
-}
-
-async fn handle_events(app: &mut App) -> Result<Option<AppAction>, AppError> {
-    if event::poll(Duration::from_millis(EVENT_POLL_INTERVAL_MS))? {
-        let ev = event::read()?;
-        if let Event::Key(key) = ev {
-            if key.kind != KeyEventKind::Press {
-                return Ok(None);
-            }
-            // Handle menu navigation
-            if app.view_mode == ViewMode::Menu {
-                match key.code {
-                    KeyCode::Up | KeyCode::Char('k') => {
-                        if app.selected_menu_item > 0 {
-                            app.selected_menu_item -= 1;
-                            app.character_count = MENU_OPTIONS[app.selected_menu_item];
-                        }
-                    }
-                    KeyCode::Down | KeyCode::Char('j') => {
-                        if app.selected_menu_item < MENU_OPTIONS.len() - 1 {
-                            app.selected_menu_item += 1;
-                            app.character_count = MENU_OPTIONS[app.selected_menu_item];
-                        }
-                    }
-                    KeyCode::Enter => {
-                        app.character_count = MENU_OPTIONS[app.selected_menu_item];
-                        return Ok(Some(AppAction::StartTraining));
-                    }
-                    KeyCode::Char('r') => {
-                        // Show report from menu
-                        app.view_mode = ViewMode::Report;
-                        app.status_message = "Report. Press 'r' to close.".to_string();
-                    }
-                    KeyCode::Char('h') => {
-                        // Show help from menu
-                        app.view_mode = ViewMode::Help;
-                        app.status_message = "Help. Press 'h' to close.".to_string();
-                    }
-                    KeyCode::Char('q') => {
-                        app.should_quit = true;
-                    }
-                    _ => {}
-                }
-                return Ok(None);
-            }
-
-            if app.is_editing {
-                // Check for Ctrl+S to submit (Shift+Enter doesn't work in most terminals)
-                if key.code == KeyCode::Char('s') && key.modifiers.contains(KeyModifiers::CONTROL) {
-                    // Ctrl+S: Submit for evaluation
-                    let content = app.text_area_state.value().to_string();
-                    if !content.trim().is_empty() {
-                        app.is_editing = false;
-                        app.text_area_state.focus.set(false); // Disable focus
-                        return Ok(Some(AppAction::Evaluate));
-                    }
-                } else if key.code == KeyCode::Esc {
-                    app.is_editing = false;
-                    app.text_area_state.focus.set(false); // Disable focus
-                    app.status_message = "Normal Mode. Press 'i' to edit.".to_string();
-                } else {
-                    // Pass all other input to rat-text TextArea
-                    // Use HandleEvent trait
-                    let _ = app.text_area_state.handle(&ev, rat_text::event::Regular);
-                }
-            } else {
-                // Handle Report view
-                if app.view_mode == ViewMode::Report {
-                    match key.code {
-                        KeyCode::Char('r') => {
-                            app.return_from_aux_view();
-                        }
-                        KeyCode::Char('q') => {
-                            app.should_quit = true;
-                        }
-                        _ => {}
-                    }
-                    return Ok(None);
-                }
-
-                // Handle Help view
-                if app.view_mode == ViewMode::Help {
-                    match key.code {
-                        KeyCode::Char('h') => {
-                            app.return_from_aux_view();
-                            app.help_scroll = 0;
-                        }
-                        KeyCode::Down | KeyCode::Char('j') => {
-                            app.help_scroll = app.help_scroll.saturating_add(1);
-                        }
-                        KeyCode::Up | KeyCode::Char('k') => {
-                            app.help_scroll = app.help_scroll.saturating_sub(1);
-                        }
-                        KeyCode::Char('q') => {
-                            app.should_quit = true;
-                        }
-                        _ => {}
-                    }
-                    return Ok(None);
-                }
-
-                match key.code {
-                    KeyCode::Char('i') | KeyCode::Enter => {
-                        if !app.show_evaluation_overlay && app.view_mode == ViewMode::Normal {
-                            app.is_editing = true;
-                            app.text_area_state.focus.set(true); // Enable focus for input!
-                            app.text_area_state.scroll_cursor_to_visible(); // Keep cursor viewport sane
-                            app.status_message = "Editing Mode. Press 'Esc' to exit.".to_string();
-                        }
-                    }
-                    KeyCode::Char('e') => {
-                        // Toggle evaluation overlay (only if evaluation exists)
-                        if app.view_mode == ViewMode::Normal && !app.evaluation_text.is_empty() {
-                            app.show_evaluation_overlay = !app.show_evaluation_overlay;
-                            if app.show_evaluation_overlay {
-                                app.evaluation_overlay_scroll = 0;
-                            }
-                        }
-                    }
-                    KeyCode::Char('n') => {
-                        // Next training: close evaluation overlay and proceed
-                        if app.show_evaluation_overlay && app.view_mode == ViewMode::Normal {
-                            app.show_evaluation_overlay = false;
-                            return Ok(Some(AppAction::NextTraining));
-                        }
-                    }
-                    KeyCode::Char('r') => {
-                        // Toggle report
-                        if app.view_mode == ViewMode::Report {
-                            app.return_from_aux_view();
-                        } else {
-                            app.view_mode = ViewMode::Report;
-                            app.status_message = "Report. Press 'r' to close.".to_string();
-                        }
-                    }
-                    KeyCode::Char('h') => {
-                        // Toggle help
-                        if app.view_mode == ViewMode::Help {
-                            app.return_from_aux_view();
-                            app.help_scroll = 0;
-                        } else {
-                            app.view_mode = ViewMode::Help;
-                            app.status_message = "Help. Press 'h' to close.".to_string();
-                        }
-                    }
-                    KeyCode::Char('q') => {
-                        app.should_quit = true;
-                    }
-                    KeyCode::Down | KeyCode::Char('j') => {
-                        if app.view_mode == ViewMode::Normal {
-                            if app.show_evaluation_overlay
-                                && key.modifiers.contains(KeyModifiers::SHIFT)
-                            {
-                                // Scroll evaluation overlay with bounds checking
-                                // Calculate visible height: overlay percent of screen minus borders and headers
-                                let visible_height = (app.terminal_height * OVERLAY_SIZE_PERCENT
-                                    / 100)
-                                    .saturating_sub(4);
-                                let visible_width = (app.terminal_width * OVERLAY_SIZE_PERCENT
-                                    / 100)
-                                    .saturating_sub(2);
-                                let max_scroll = calculate_max_scroll(
-                                    &app.evaluation_text,
-                                    visible_height,
-                                    visible_width,
-                                );
-                                app.evaluation_overlay_scroll = app
-                                    .evaluation_overlay_scroll
-                                    .saturating_add(1)
-                                    .min(max_scroll);
-                            } else {
-                                // Scroll original text with bounds checking
-                                // Calculate visible height: half screen minus header and status bar
-                                let visible_height = (app.terminal_height / 2).saturating_sub(3);
-                                let visible_width = (app.terminal_width / 2).saturating_sub(2);
-                                let max_scroll = calculate_max_scroll(
-                                    &app.original_text,
-                                    visible_height,
-                                    visible_width,
-                                );
-                                app.original_text_scroll =
-                                    app.original_text_scroll.saturating_add(1).min(max_scroll);
-                            }
-                        }
-                    }
-                    KeyCode::Up | KeyCode::Char('k') => {
-                        if app.view_mode == ViewMode::Normal {
-                            if app.show_evaluation_overlay
-                                && key.modifiers.contains(KeyModifiers::SHIFT)
-                            {
-                                app.evaluation_overlay_scroll =
-                                    app.evaluation_overlay_scroll.saturating_sub(1);
-                            } else {
-                                app.original_text_scroll =
-                                    app.original_text_scroll.saturating_sub(1);
-                            }
-                        }
-                    }
-                    _ => {}
-                }
-            }
-        }
-    }
-    Ok(None)
-}
-
-/// Calculate the maximum scroll offset for given text content
-fn calculate_max_scroll(text: &str, visible_height: u16, visible_width: u16) -> u16 {
-    if visible_width == 0 {
-        return 0;
-    }
-    let paragraph = Paragraph::new(text).wrap(Wrap { trim: false });
-    let total_lines = paragraph.line_count(visible_width) as u16;
-    total_lines.saturating_sub(visible_height.saturating_sub(2)) // -2 for borders
 }
 
 /// Generate text using the API client and update app state
